requirements.txt
oss2==2.18.5
google-cloud-storage==2.16.0
click==8.1.7